trainModels
