--- oracle/database.py
@@ -1,6 +1,0 @@
-import redis
-
-database = redis.Redis()
-database.mset({"Croatia": "Zagreb", "Bahamas": "Nassau"})
-database.get("Bahamas")
-database.save()
