+++ oracle/oracle.py
@@ -7,7 +7,7 @@ import json
 import copy
 import requests
 import models
-import database
+import dataManager
 sys.path.append("../")
 import utils
 
@@ -28,7 +28,7 @@ class Pricing:
         """Calculates and returns the latest price and the txn_id where it was changed"""
         mult, txn_id = self.get_price_multiplier(utils.OpCodes.TRAIN_MODEL)
         model = models.PredictModel.create(raw_model, **kwargs)
-        dataset = database.database.get(dataset_name)
+        dataset = dataManager.database.get(dataset_name)
         return model.model_complexity * mult * dataset["size"], txn_id
 
     def calc_model_query_price(self, model_name: str):
@@ -42,7 +42,7 @@ class Pricing:
         if not self.mult_cache.get(op):
             # get price from database
 
-            self.mult_cache[op] = database.database.get("<PRICE>"+op)
+            self.mult_cache[op] = dataManager.database.get("<PRICE>"+op)
 
         return self.mult_cache[op]["mul"], self.mult_cache[op]["txn_id"]
 
@@ -53,10 +53,10 @@ class Pricing:
 
         self.mult_cache[op] = {"op": op, "mul": new_mul, "txn_id": txn["id"]}
         # Save txn_id to database
-        database.database.set("<PRICE>"+op, {"op": op, "mul": new_mul, "txn_id": txn["id"]})
+        dataManager.database.set("<PRICE>"+op, {"op": op, "mul": new_mul, "txn_id": txn["id"]})
 
 
-class ClientTransactionMonitor(utils.TransactionMonitor):
+class OracleTransactionMonitor(utils.TransactionMonitor):
 
     def process_incoming(self, txn):
         """Execute operations based on the OP code of the incoming transaction"""
@@ -66,117 +66,17 @@ class ClientTransactionMonitor(utils.TransactionMonitor):
 
         match op:
             case utils.OpCodes.UP_DATASET:
-                handler = LocalDataHandler(args[1])
-                save_dataset(handler, args[0])
+                handler = dataManager.LocalDataHandler(args[1])
+                return dataManager.save_dataset(handler, args[0])
             case utils.OpCodes.QUERY_MODEL:
-                ...
+                model = models.get_trained_model(args[0])
+                return model(args[1])
             case utils.OpCodes.UPDATE_PRICE:
                 ...
-
-
-class DataHandler:
-
-    SAVE_MODE = 0
-    LOAD_MODE = 1
-
-    def __init__(self, dataset_name):
-        self.dataset_name = dataset_name
-        self.env = ""
-
-    @abc.abstractmethod
-    def start(self, mode: int):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def save_chunk(self, data: bytes):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def save_all(self, data: bytes):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def load_chunk(self, chunk_size=1024):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def load_all(self):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def finish(self):
-        raise NotImplementedError()
-
-
-class LocalDataHandler(DataHandler):
-
-    def __init__(self, dataset_name):
-        super().__init__(dataset_name)
-        self.env = "local"
-        self.dataset_name = dataset_name
-        self.file = None
-        self.file_path = f"data/{dataset_name}"
-        self.mode = None
-
-    def start(self, mode: int):
-        self.mode = mode
-        if mode == self.SAVE_MODE:
-            if os.path.isfile(self.file_path):
-                raise FileExistsError()
-            self.file = open(self.file_path, "ab")
-        elif mode == self.LOAD_MODE:
-            self.file = open(self.file_path, "rb")
-
-    def save_chunk(self, data: bytes):
-        if self.mode != self.SAVE_MODE:
-            raise AttributeError("Cannot perform save operation when not in save mode!")
-        self.file.write(data)
-
-    def save_all(self, data: bytes):
-        if self.mode != self.SAVE_MODE:
-            raise AttributeError("Cannot perform save operation when not in save mode!")
-        self.file.write(data)
-
-    def load_chunk(self, chunk_size=1024):
-        if self.mode != self.LOAD_MODE:
-            raise AttributeError("Cannot perform load operation when not in load mode!")
-        return self.file.read(chunk_size)
-
-    def load_all(self):
-        if self.mode != self.LOAD_MODE:
-            raise AttributeError("Cannot perform load operation when not in load mode!")
-        return self.file.read()
-
-    def finish(self):
-        self.file.close()
-        self.file = None
-        self.mode = None
-
-
-def save_dataset(handler: DataHandler, link: str):
-    handler.start(handler.SAVE_MODE)
-    size = 0
-    with requests.get(link, stream=True) as r:
-        r.raise_for_status()
-        for chunk in r.iter_content(chunk_size=8192):
-            size += len(chunk)
-            handler.save_chunk(chunk)
-
-    handler.finish()
-    database.database.set("<DS>"+handler.dataset_name, {"env": handler.env, "size": size})
-
-
-def load_dataset(handler: DataHandler):
-    handler.start(handler.LOAD_MODE)
-    data = b''
-    chunk = handler.load_chunk()
-    while chunk:
-        data += chunk
-        chunk = handler.load_chunk()
-
-    handler.finish()
-    dataset = database.database.get("<DS>"+handler.dataset_name)
-    return {**dataset, "data": data}
+            case utils.OpCodes.TRAIN_MODEL:
+                model = models.PredictModel.create(args[0], )
+                handler = dataManager.LocalDataHandler(args[1])
+                dataset = dataManager.load_dataset(handler)
 
 
 app = Flask(__name__)
